adding preview for the local fonts
