--- src/fonts-page-layout/fonts-page-layout.css
@@ -1,18 +1,0 @@
-.fonts-page-layout {
-	display: grid;
-	grid-template-columns: 1fr 250px;
-	margin: 0;
-	padding: 0;
-}
-
-.fonts-page-layout > main {
-	padding: 2rem;
-}
-
-#wpfooter {
-	display: none;
-}
-
-#wpbody-content {
-	padding: 0;
-}
