--- src/fonts-page-layout/index.js
@@ -1,7 +1,0 @@
-import './fonts-page-layout.css';
-
-function FontsPageLayout( props ) {
-	return <div className="fonts-page-layout">{ props.children }</div>;
-}
-
-export default FontsPageLayout;
