+++ src/local-fonts/local-fonts.css
@@ -1,3 +1,13 @@
+.layout {
+	display: grid;
+	grid-template-columns: 350px 1fr;
+}
+
+.layout > div,
+.layout main {
+	padding: 2rem;
+}
+
 .form-group {
 	margin-bottom: 1rem;
 	max-width: 300px;
@@ -5,4 +15,5 @@
 
 .form-group input {
 	width: 100%;
+	column-gap: 2rem;
 }
